Generalized Student/TeacherProfile to Profile
