--- app/src/main/java/com/example/quizapp/ui/home/base/BaseProfileFragment.kt
@@ -1,20 +1,0 @@
-package com.example.quizapp.ui.home.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.quizapp.databinding.FragmentProfileBinding
-
-open class BaseProfileFragment() : Fragment() {
-    protected lateinit var binding: FragmentProfileBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentProfileBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-}
