+++ app/src/main/java/com/example/quizapp/ui/home/teacher/TeacherHomeViewModel.kt
@@ -2,7 +2,7 @@ package com.example.quizapp.ui.home.teacher
 
 import com.example.quizapp.core.service.AuthService
 import com.example.quizapp.data.repo.QuizRepo
-import com.example.quizapp.ui.base.BaseViewModel
+import com.example.quizapp.ui.home.base.BaseHomeViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -10,11 +10,8 @@ import javax.inject.Inject
 @HiltViewModel
 class TeacherHomeViewModel @Inject constructor(
     private val repo: QuizRepo,
-    private val authService: AuthService
-) : BaseViewModel() {
-    fun getUserPhoto() = authService.getUserPhoto()
-    fun logout() = authService.logout()
-
+    authService: AuthService
+) : BaseHomeViewModel(authService) {
 //    private val _state = MutableStateFlow(HomeState())
 //    val state = _state.asStateFlow()
 //
